backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
pydantic==2.10.4

# ML / Deep Learning
torch>=2.1.0
torchvision>=0.16.0
smplx==0.1.28
numpy>=1.24.0
scipy>=1.11.0

# Image processing
Pillow>=10.0.0
opencv-python-headless>=4.8.0

# 3D mesh
trimesh>=4.0.0

# Deployment
modal>=0.67.0

# Utils
python-dotenv>=1.0.0
